Refactor logging and URL handling in video analysis
Modified Program.cs to comment out the line setting the VideoFrame URL in the response and added a log statement to log the VideoResponse object. Added a ToString method to the VideoResponse class in VideoResponse.cs to provide a string representation of the object. Updated Home.razor to inject ILogger<Program> instead of ILogger<Index>. Changed AnalyzeVideoAsync method in VideoAnalysisApiClient.cs to accept ILogger<Program> instead of ILogger<Index>, set the VideoFrame property of the VideoResponse object to include the base address of the HTTP client, and log the VideoResponse object.